refactor percentile_feature_extractor
